+++ bob/src/api/http.rs
@@ -111,6 +111,11 @@ pub(crate) struct VersionInfo {
     pearl_version: Version,
 }
 
+#[derive(Debug, Serialize)]
+pub(crate) struct NodeConfiguration {
+    blob_file_name_prefix: String,
+}
+
 pub(crate) fn spawn(bob: BobServer, address: IpAddr, port: u16) {
     let routes = routes![
         status,
@@ -121,7 +126,9 @@ pub(crate) fn spawn(bob: BobServer, address: IpAddr, port: u16) {
         partition_by_id,
         change_partition_state,
         delete_partition,
+        get_node_configuration,
         alien,
+        detach_alien_partitions,
         get_alien_directory,
         remount_vdisks_group,
         start_all_disk_controllers,
@@ -288,12 +295,12 @@ async fn nodes(bob: &State<BobServer>) -> Json<Vec<Node>> {
 #[get("/disks/list")]
 async fn disks_list(bob: &State<BobServer>) -> Result<Json<Vec<DiskState>>, StatusExt> {
     let backend = bob.grinder().backend().inner();
-    let (dcs, adc) = backend
+    let (dcs, alien_disk_controller) = backend
         .disk_controllers()
         .ok_or_else(not_acceptable_backend)?;
 
     let mut disks = Vec::new();
-    for dc in dcs.iter().chain(std::iter::once(&adc)) {
+    for dc in dcs.iter().chain(std::iter::once(&alien_disk_controller)) {
         let disk_path = dc.disk();
         disks.push(DiskState {
             name: disk_path.name().to_owned(),
@@ -313,6 +320,15 @@ async fn distribution_function(bob: &State<BobServer>) -> Json<DistrFunc> {
     })
 }
 
+#[get("/configuration")]
+async fn get_node_configuration(bob: &State<BobServer>) -> Json<NodeConfiguration> {
+    let grinder = bob.grinder();
+    let config = grinder.node_config();
+    Json(NodeConfiguration {
+        blob_file_name_prefix: config.pearl().blob_file_name_prefix().to_owned(),
+    })
+}
+
 #[post("/disks/<disk_name>/stop")]
 async fn stop_all_disk_controllers(
     bob: &State<BobServer>,
@@ -320,11 +336,11 @@ async fn stop_all_disk_controllers(
 ) -> Result<StatusExt, StatusExt> {
     use futures::stream::{FuturesUnordered, StreamExt};
     let backend = bob.grinder().backend().inner();
-    let (dcs, adc) = backend
+    let (dcs, alien_disk_controller) = backend
         .disk_controllers()
         .ok_or_else(not_acceptable_backend)?;
     dcs.iter()
-        .chain(std::iter::once(&adc))
+        .chain(std::iter::once(&alien_disk_controller))
         .filter(|dc| dc.disk().name() == disk_name)
         .map(|dc| dc.stop())
         .collect::<FuturesUnordered<_>>()
@@ -344,12 +360,12 @@ async fn start_all_disk_controllers(
 ) -> Result<StatusExt, StatusExt> {
     use futures::stream::{FuturesUnordered, StreamExt};
     let backend = bob.grinder().backend().inner();
-    let (dcs, adc) = backend
+    let (dcs, alien_disk_controller) = backend
         .disk_controllers()
         .ok_or_else(not_acceptable_backend)?;
     let target_dcs = dcs
         .iter()
-        .chain(std::iter::once(&adc))
+        .chain(std::iter::once(&alien_disk_controller))
         .filter(|dc| dc.disk().name() == disk_name)
         .map(|dc| dc.run())
         .collect::<FuturesUnordered<_>>();
@@ -568,13 +584,23 @@ async fn alien(_bob: &State<BobServer>) -> &'static str {
     "alien"
 }
 
+#[post("/alien/detach")]
+async fn detach_alien_partitions(bob: &State<BobServer>) -> Result<StatusExt, StatusExt> {
+    let backend = bob.grinder().backend().inner();
+    let (_, alien_disk_controller) = backend
+        .disk_controllers()
+        .ok_or_else(not_acceptable_backend)?;
+    alien_disk_controller.detach_all().await?;
+    Ok(StatusExt::new(Status::Ok, true, String::default()))
+}
+
 #[get("/alien/dir")]
 async fn get_alien_directory(bob: &State<BobServer>) -> Result<Json<Dir>, StatusExt> {
     let backend = bob.grinder().backend().inner();
-    let (_, adc) = backend
+    let (_, alien_disk_controller) = backend
         .disk_controllers()
         .ok_or_else(not_acceptable_backend)?;
-    let path = PathBuf::from(adc.disk().path());
+    let path = PathBuf::from(alien_disk_controller.disk().path());
     let dir = create_directory(&path).await?;
     Ok(Json(dir))
 }